refactor: extract TrackerExportSupport + siblings (#649 slice B FINAL)
Second and final slice of #649 (ExportService.swift decomposition). Extracts
all 5 support types + the TrackerExportSupport extension into a new
TrackerExportSupport.swift:
- TrackerIssueCandidate (struct)
- TrackerExportSupport (enum)
- ExportRetryConfiguration (struct)
- ExportCreateOutcome (enum)
- TrackerExportRetryContext (struct)
- extension TrackerExportSupport (retry logic)

These 5 types + extension form one cohesive TrackerExportSupport domain
(retry/candidate/outcome types used by ExportService's tracker-export flow).
Bundling them into one file per intent-first coherence.

Byte-identical move; SHA-256 329187372d5a5c5f09b06986d1bbd914545e458cc6a0bf09767fc5ac5394b467
verified against origin/main:ExportService.swift lines 111-350.

Reduces ExportService.swift from 350 lines to 109 lines. After all #649
slices (A: SimilarIssueReviewService, B: TrackerExportSupport), ExportService.swift
went from 666 lines to 109 lines (~84% reduction), containing only the
primary ExportService actor.

Tests: 667 (unchanged).
